Refactor termvectors
