+++ specification/_global/termvectors/types.ts
@@ -17,8 +17,39 @@
  * under the License.
  */
 
-import { double, integer } from '@_types/Numeric'
-import { Token } from './Token'
+import { Dictionary } from "@spec_utils/Dictionary"
+import { Id, IndexName, Field, VersionNumber } from "@_types/common"
+import { integer, long, double } from "@_types/Numeric"
+
+export class FieldStatistics {
+  doc_count: integer
+  sum_doc_freq: long
+  sum_ttf: long
+}
+
+export class TermVector {
+  field_statistics: FieldStatistics
+  terms: Dictionary<string, TermVectorTerm>
+}
+
+export class TermVectorFilter {
+  max_doc_freq?: integer
+  max_num_terms?: integer
+  max_term_freq?: integer
+  max_word_length?: integer
+  min_doc_freq?: integer
+  min_term_freq?: integer
+  min_word_length?: integer
+}
+
+export class TermVectorsResult {
+  found: boolean
+  id: Id
+  index: IndexName
+  term_vectors: Dictionary<Field, TermVector>
+  took: long
+  version: VersionNumber
+}
 
 export class TermVectorTerm {
   doc_freq?: integer
@@ -26,4 +57,11 @@ export class TermVectorTerm {
   term_freq: integer
   tokens: Token[]
   ttf?: integer
+}
+
+export class Token {
+  end_offset?: integer
+  payload?: string
+  position: integer
+  start_offset?: integer
 }
